used options menu to create overflow menu and got rid of the menu handler

+++ SharedModule/src/main/java/com/soogbad/sharedmodule/ui/ItemActionBar.java
@@ -3,10 +3,9 @@ package com.soogbad.sharedmodule.ui;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.util.AttributeSet;
-import android.view.View;
+import android.view.MenuItem;
 import android.view.LayoutInflater;
 import android.widget.EditText;
-import android.widget.ImageButton;
 import android.widget.Toast;
 
 import androidx.constraintlayout.widget.ConstraintLayout;
@@ -16,13 +15,10 @@ import com.soogbad.sharedmodule.core.ItemsManager;
 import com.soogbad.sharedmodule.R;
 import com.soogbad.sharedmodule.core.Utility;
 
-import java.util.Map;
-
 public class ItemActionBar extends ConstraintLayout {
 
     private final EditText titleEditText;
     public EditText getTitleEditText() { return titleEditText; }
-    protected final ImageButton overflowMenuButton;
 
     private ItemLayout itemLayout;
     private Item<?> item;
@@ -30,45 +26,42 @@ public class ItemActionBar extends ConstraintLayout {
     public ItemActionBar(Context context, AttributeSet attrs) {
         super(context, attrs);
         LayoutInflater.from(context).inflate(R.layout.item_action_bar_content, this, true);
-        titleEditText = findViewById(R.id.titleEditText); overflowMenuButton = findViewById(R.id.overflowMenuButton);
+        titleEditText = findViewById(R.id.titleEditText);
     }
 
     public void init(ItemLayout itemLayout, Item<?> item, boolean readOnly) {
         this.itemLayout = itemLayout; this.item = item;
         titleEditText.setText(item.Title);
-        overflowMenuButton.setOnClickListener(readOnly ? this::showRecycleBinOverflowMenu : this::showOverflowMenu);
         if(readOnly) {
             titleEditText.setFocusable(false); titleEditText.setFocusableInTouchMode(false); titleEditText.setCursorVisible(false);
         }
     }
 
-    private void showOverflowMenu(View view) {
-        new ItemMenuHandler(getContext(), Map.ofEntries(
-            Map.entry(R.id.action_delete, () -> {
-                itemLayout.delete();
-                Utility.getActivity(getContext()).finish();
-            }),
-            Map.entry(R.id.action_copy_uuid, () ->
-                Utility.getAppUtility(getContext()).copyItemUuid(getContext(), item)
-            )
-        )).showOverflowMenu(view, R.menu.item_menu);
-    }
-
-    private void showRecycleBinOverflowMenu(View view) {
+    public boolean onOptionsItemSelected(MenuItem menuItem) {
         Context context = getContext();
         ItemsManager<?, ?> itemsManager = Utility.getItemsManager(context);
-        new ItemMenuHandler(context, Map.ofEntries(
-            Map.entry(R.id.action_restore, () -> {
-                itemsManager.restoreRecycleBinItem(item.UUID);
-                Toast.makeText(context, "Item restored", Toast.LENGTH_SHORT).show();
-                overflowMenuButton.setOnClickListener(this::showOverflowMenu);
-                titleEditText.setFocusable(true); titleEditText.setFocusableInTouchMode(true); titleEditText.setCursorVisible(true);
-                itemLayout.disableReadOnly(); ((ItemActivity)Utility.getActivity(context)).disablePreviewMode();
-            }),
-            Map.entry(R.id.action_delete_permanently, () ->
-                new AlertDialog.Builder(context).setTitle("Delete Permanently").setMessage("Are you sure you want to permanently delete this item?")
-                        .setPositiveButton("Delete", (dialog, which) -> { itemsManager.permanentlyDeleteRecycleBinItem(item.UUID); Utility.getActivity(context).finish(); }).setNegativeButton("Cancel", null).show())
-        )).showOverflowMenu(view, R.menu.recycle_bin_item_menu);
+        if(menuItem.getItemId() == R.id.action_copy_uuid) {
+            Utility.getAppUtility(getContext()).copyItemUuid(getContext(), item);
+            return true;
+        }
+        else if(menuItem.getItemId() == R.id.action_delete) {
+            itemLayout.delete();
+            Utility.getActivity(getContext()).finish();
+            return true;
+        }
+        else if(menuItem.getItemId() == R.id.action_restore) {
+            itemsManager.restoreRecycleBinItem(item.UUID);
+            Toast.makeText(context, "Item restored", Toast.LENGTH_SHORT).show();
+            titleEditText.setFocusable(true); titleEditText.setFocusableInTouchMode(true); titleEditText.setCursorVisible(true);
+            itemLayout.disableReadOnly(); ((ItemActivity)Utility.getActivity(context)).disablePreviewMode();
+            return true;
+        }
+        else if(menuItem.getItemId() == R.id.action_delete_permanently) {
+            new AlertDialog.Builder(context).setTitle("Delete Permanently").setMessage("Are you sure you want to permanently delete this item?")
+                    .setPositiveButton("Delete", (dialog, which) -> { itemsManager.permanentlyDeleteRecycleBinItem(item.UUID); Utility.getActivity(context).finish(); }).setNegativeButton("Cancel", null).show();
+            return true;
+        }
+        return false;
     }
 
 }
